Clean up network code

+++ main_shared/aistruct.h
@@ -140,7 +140,7 @@ typedef struct ai_static {
 	short			hide_segment;				//	Segment to go to for hiding.
 	short			hide_index;					//	Index in Path_seg_points
 	short			path_length;				//	Length of hide path.
-	int8_t			cur_path_index;			//	Current index in path.
+	short			cur_path_index;			//	Current index in path.
 	int8_t			dying_sound_playing;		//	!0 if this robot is playing its dying sound.
 
 short			follow_path_start_seg;	//	Start segment for robot which follows path.
